bpo-32032: Test both implementations of module-level pickle API.
