+++ Lib/test/test_pickle.py
@@ -26,8 +26,12 @@ except ImportError:
     has_c_implementation = False
 
 
-class PickleTests(AbstractPickleModuleTests):
-    pass
+class PyPickleTests(AbstractPickleModuleTests):
+    dump = staticmethod(pickle._dump)
+    dumps = staticmethod(pickle._dumps)
+    load = staticmethod(pickle._load)
+    loads = staticmethod(pickle._loads)
+    Pickler = pickle._Pickler
 
 
 class PyUnpicklerTests(AbstractUnpickleTests):
@@ -136,6 +140,21 @@ class PyChainDispatchTableTests(AbstractDispatchTableTests):
 
 
 if has_c_implementation:
+    class CPickleTests(AbstractPickleModuleTests):
+        from _pickle import dump, dumps, load, loads, Pickler
+
+        def test_bad_init(self):
+            # Test issue3664 (pickle can segfault from a badly initialized Pickler).
+            # Override initialization without calling __init__() of the superclass.
+            class BadPickler(pickle.Pickler):
+                def __init__(self): pass
+
+            class BadUnpickler(pickle.Unpickler):
+                def __init__(self): pass
+
+            self.assertRaises(pickle.PicklingError, BadPickler().dump, 0)
+            self.assertRaises(pickle.UnpicklingError, BadUnpickler().load)
+
     class CUnpicklerTests(PyUnpicklerTests):
         unpickler = _pickle.Unpickler
         bad_stack_errors = (pickle.UnpicklingError,)
@@ -426,12 +445,12 @@ class CompatPickleTests(unittest.TestCase):
 
 
 def test_main():
-    tests = [PickleTests, PyUnpicklerTests, PyPicklerTests,
+    tests = [PyPickleTests, PyUnpicklerTests, PyPicklerTests,
              PyPersPicklerTests, PyIdPersPicklerTests,
              PyDispatchTableTests, PyChainDispatchTableTests,
              CompatPickleTests]
     if has_c_implementation:
-        tests.extend([CUnpicklerTests, CPicklerTests,
+        tests.extend([CPickleTests, CUnpicklerTests, CPicklerTests,
                       CPersPicklerTests, CIdPersPicklerTests,
                       CDumpPickle_LoadPickle, DumpPickle_CLoadPickle,
                       PyPicklerUnpicklerObjectTests,
